# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Rathinam Crackers Business Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /v1/dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary widgets
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /v1/dashboard/sales-by-channel:
    get:
      operationId: getSalesByChannel
      tags: [dashboard]
      summary: Get sales grouped by channel (chart data)
      security:
        - bearerAuth: []
      parameters:
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Sales by channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesByChannelResponse"
  /v1/dashboard/top-products:
    get:
      operationId: getTopProducts
      tags: [dashboard]
      summary: Get top selling products
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopProductsResponse"
  /v1/dashboard/agent-performance:
    get:
      operationId: getAgentPerformance
      tags: [dashboard]
      summary: Get agent performance this month
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Agent performance
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentPerformanceResponse"
  /v1/dashboard/by-location:
    get:
      operationId: getDashboardByLocation
      tags: [dashboard]
      summary: Per-location sales, stock and shift metrics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Location metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardByLocationResponse"
  /v1/dashboard/daily-sales:
    get:
      operationId: getDailySales
      tags: [dashboard]
      summary: Daily sales time series for last N days
      security:
        - bearerAuth: []
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 30
        - name: locationId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Daily sales series
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailySalesResponse"
  /v1/dashboard/business-overview:
    get:
      operationId: getBusinessOverview
      tags: [dashboard]
      summary: Business-wide KPIs (MTD, YTD, growth, payment mix, top customers, GST)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Business overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessOverviewResponse"

  # ============================================================
  # PRODUCTS
  # ============================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DashboardSummary:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            todaySales:
              type: number
            todayInvoices:
              type: integer
            outstandingTotal:
              type: number
            lowStockCount:
              type: integer
            activeTransfers:
              type: integer
            monthSales:
              type: number
            monthGrowth:
              type: number
    SalesByChannelItem:
      type: object
      properties:
        channel:
          type: string
        revenue:
          type: number
        count:
          type: integer
    SalesByChannelResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: "#/components/schemas/SalesByChannelItem"
    TopProductsResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              productName:
                type: string
              totalQty:
                type: integer
              totalRevenue:
                type: number
    AgentPerformanceResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            type: object
            properties:
              agentId:
                type: string
              agentName:
                type: string
              sales:
                type: number
              target:
                type: number
              achievement:
                type: number
    DashboardByLocationItem:
      type: object
      properties:
        locationId: { type: string }
        locationName: { type: string }
        type: { type: string }
        city: { type: string, nullable: true }
        todaySales: { type: number }
        todayInvoices: { type: integer }
        monthSales: { type: number }
        monthInvoices: { type: integer }
        averageOrderValue: { type: number }
        lowStockCount: { type: integer }
        openShifts: { type: integer }
    DashboardByLocationResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: array
          items:
            $ref: "#/components/schemas/DashboardByLocationItem"
    DailySalesPoint:
      type: object
      properties:
        date: { type: string }
        revenue: { type: number }
        count: { type: integer }
    DailySalesResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: array
          items:
            $ref: "#/components/schemas/DailySalesPoint"
    BusinessOverviewResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            mtdRevenue: { type: number }
            mtdInvoices: { type: integer }
            prevMonthRevenue: { type: number }
            monthGrowth: { type: number }
            ytdRevenue: { type: number }
            ytdInvoices: { type: integer }
            averageOrderValue: { type: number }
            activeCustomers: { type: integer }
            activeProducts: { type: integer }
            gstCollectedMtd: { type: number }
            gstBreakdown:
              type: object
              properties:
                cgst: { type: number }
                sgst: { type: number }
                igst: { type: number }
            paymentModeMix:
              type: array
              items:
                type: object
                properties:
                  mode: { type: string }
                  revenue: { type: number }
                  count: { type: integer }
            topCustomers:
              type: array
              items:
                type: object
                properties:
                  customerId: { type: string, nullable: true }
                  customerName: { type: string }
                  revenue: { type: number }
                  orders: { type: integer }

    # ============================================================
    # SITE CONTENT (CMS)
    # ============================================================
...